Move useProductAndPlans as a shared hook.
