+++ client/a8c-for-agencies/sections/marketplace/products-overview/product-listing/index.tsx
@@ -13,10 +13,10 @@ import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import FilterSearch from '../../../../components/filter-search';
 import { ShoppingCartContext } from '../../context';
+import useProductAndPlans from '../../hooks/use-product-and-plans';
 import ListingSection from '../../listing-section';
 import MultiProductCard from '../multi-product-card';
 import ProductCard from '../product-card';
-import useProductAndPlans from './hooks/use-product-and-plans';
 import { getSupportedBundleSizes, useProductBundleSize } from './hooks/use-product-bundle-size';
 import useSubmitForm from './hooks/use-submit-form';
 import VolumePriceSelector from './volume-price-selector';
